# docs/swagger.yaml
basePath: /
definitions:
...
  ListUsersRow:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
  Role:
    enum:
    - editor
    - reviewer
    - admin
    type: string
    x-enum-comments:
      RoleAdmin: + manages users
      RoleEditor: creates and edits drugs, packages and leaflets
      RoleReviewer: + marks leaflets as reviewed (pharmacist)
    x-enum-descriptions:
    - creates and edits drugs, packages and leaflets
    - + marks leaflets as reviewed (pharmacist)
    - + manages users
    x-enum-varnames:
    - RoleEditor
    - RoleReviewer
    - RoleAdmin
...
  createUserRequest:
    properties:
      email:
        example: ana@farmacia.com
        type: string
      name:
        example: Ana Souza
        type: string
      password:
        example: uma-senha-forte
        type: string
      role:
        allOf:
        - $ref: '#/definitions/Role'
        enum:
        - editor
        - reviewer
        - admin
        example: reviewer
    type: object
  idResponse:
    properties:
      id:
        example: 1
        type: integer
    type: object
...
  setPasswordRequest:
    properties:
      password:
        example: nova-senha-forte
        type: string
    type: object
...
  updateUserRequest:
    properties:
      active:
        example: true
        type: boolean
      email:
        example: ana@farmacia.com
        type: string
      name:
        example: Ana Souza
        type: string
      role:
        allOf:
        - $ref: '#/definitions/Role'
        enum:
        - editor
        - reviewer
        - admin
        example: reviewer
    type: object
info:
  contact: {}
  description: |-
    API that reads the barcode on a medicine box and returns the simplified leaflet.
    Public route: GET /drugs/ean/{ean}. All others require login (POST /auth/login sets a session cookie).
  title: PharmaCode API
  version: "1.0"
paths:
...
  /users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ListUsersRow'
            type: array
        "401":
          description: authentication required
          schema:
            type: string
        "403":
          description: insufficient permissions
          schema:
            type: string
      summary: List admin panel users (admin)
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: User data (password of 8 to 72 characters)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/createUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/idResponse'
        "400":
          description: invalid data
          schema:
            type: string
        "409":
          description: email already registered
          schema:
            type: string
      summary: Create an admin panel user (admin)
      tags:
      - users
  /users/{id}:
    put:
      consumes:
      - application/json
      description: Deactivating a user or changing their role ends their sessions.
        The last active admin can't be demoted or deactivated.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/updateUserRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: invalid data
          schema:
            type: string
        "404":
          description: user not found
          schema:
            type: string
        "409":
          description: email already registered / cannot remove the last active admin
          schema:
            type: string
      summary: Update name, email, role and status (admin)
      tags:
      - users
  /users/{id}/password:
    put:
      consumes:
      - application/json
      description: Ends all of the user's sessions.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: New password (8 to 72 characters)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/setPasswordRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: weak password
          schema:
            type: string
        "404":
          description: user not found
          schema:
            type: string
      summary: Set a new password for a user (admin)
      tags:
      - users
swagger: "2.0"